integration-tests: Re-structure connectors tests
