+++ runtime/integration-tests/src/connectors/pallet/development/setup.rs
@@ -1,4 +1,4 @@
-// Copyright 2023 Centrifuge Foundation (centrifuge.io).
+// Copyright 2021 Centrifuge Foundation (centrifuge.io).
 //
 // This file is part of the Centrifuge chain project.
 // Centrifuge is free software: you can redistribute it and/or modify
@@ -10,20 +10,17 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 
+pub use altair_runtime::{AccountId, CurrencyId, Runtime, RuntimeOrigin, System};
 use cfg_primitives::{currency_decimals, parachains, Balance};
-use cfg_types::{
-	domain_address::Domain,
-	tokens::{CurrencyId, CustomMetadata},
-};
+use cfg_types::{domain_address::Domain, tokens::CustomMetadata};
 use frame_support::traits::GenesisBuild;
 use orml_traits::asset_registry::AssetMetadata;
-
-use crate::chain::centrifuge::{AccountId, Runtime, RuntimeOrigin, System};
 
 /// Accounts
 pub const ALICE: [u8; 32] = [4u8; 32];
 pub const BOB: [u8; 32] = [5u8; 32];
 pub const CHARLIE: [u8; 32] = [6u8; 32];
+
 pub const TEST_DOMAIN: Domain = Domain::EVM(1284);
 
 /// A PARA ID used for a sibling parachain emulating Moonbeam.
@@ -110,4 +107,18 @@ pub fn cfg(amount: Balance) -> Balance {
 
 pub fn dollar(decimals: u32) -> Balance {
 	10u128.saturating_pow(decimals)
+}
+
+pub fn moonbeam_account() -> AccountId {
+	parachain_account(PARA_ID_MOONBEAM)
+}
+
+pub fn centrifuge_account() -> AccountId {
+	parachain_account(parachains::polkadot::centrifuge::ID)
+}
+
+fn parachain_account(id: u32) -> AccountId {
+	use sp_runtime::traits::AccountIdConversion;
+
+	polkadot_parachain::primitives::Sibling::from(id).into_account_truncating()
 }
