+++ runtime/integration-tests/src/connectors/pallet/kusama/tests/currency_id_convert.rs
@@ -51,8 +51,7 @@ use xcm_emulator::TestExt;
 use xcm_executor::traits::Convert as C1;
 
 use crate::{
-	utils::AUSD_CURRENCY_ID,
-	xcm::kusama::{
+	connectors::pallet::kusama::{
 		setup::{
 			air, altair_account, ausd, foreign, karura_account, ksm, sibling_account, ALICE, BOB,
 			KSM_ASSET_ID, PARA_ID_SIBLING,
@@ -60,6 +59,7 @@ use crate::{
 		test_net::{Altair, Karura, KusamaNet, Sibling, TestNet},
 		tests::{register_air, register_ausd, register_ksm},
 	},
+	utils::AUSD_CURRENCY_ID,
 };
 
 #[test]
