+++ runtime/integration-tests/src/connectors/pallet/polkadot/tests/currency_id_convert.rs
@@ -51,8 +51,7 @@ use xcm_executor::traits::Convert as C1;
 
 use super::register_dot;
 use crate::{
-	utils::AUSD_CURRENCY_ID,
-	xcm::polkadot::{
+	connectors::pallet::polkadot::{
 		setup::{
 			acala_account, ausd, centrifuge_account, cfg, dot, foreign, sibling_account, ALICE,
 			BOB, DOT_ASSET_ID, NO_XCM_ASSET_ID, PARA_ID_SIBLING,
@@ -60,6 +59,7 @@ use crate::{
 		test_net::{Acala, Centrifuge, PolkadotNet, Sibling, TestNet},
 		tests::{register_ausd, register_cfg, register_cfg_v2, register_no_xcm_token},
 	},
+	utils::AUSD_CURRENCY_ID,
 };
 
 #[test]
